# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zerofy API — AI Climate Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type CarbonInputCarType = typeof CarbonInputCarType[keyof typeof CarbonInputCarType];


export const CarbonInputCarType = {
  petrol: 'petrol',
  diesel: 'diesel',
  ev: 'ev',
  none: 'none',
} as const;

export type CarbonInputEnergySourceMix = {
  /** Percentage coal 0-100 */
  coal: number;
  /** Percentage gas 0-100 */
  gas: number;
  /** Percentage renewables 0-100 */
  renewables: number;
};

export type CarbonInputDietType = typeof CarbonInputDietType[keyof typeof CarbonInputDietType];


export const CarbonInputDietType = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  omnivore: 'omnivore',
  heavy_meat: 'heavy_meat',
} as const;

export type CarbonInputFlightType = typeof CarbonInputFlightType[keyof typeof CarbonInputFlightType];


export const CarbonInputFlightType = {
  short: 'short',
  long: 'long',
  mixed: 'mixed',
} as const;

export interface CarbonInput {
  /** Weekly km driven by car */
  weeklyKmDriven: number;
  carType: CarbonInputCarType;
  /** Monthly electricity usage in kWh */
  monthlyKwh: number;
  energySourceMix: CarbonInputEnergySourceMix;
  dietType: CarbonInputDietType;
  flightsPerYear: number;
  flightType: CarbonInputFlightType;
  monthlyShoppingSpend: number;
  /** User region for grid emission intensity */
  region: string;
}

export interface EmissionBreakdown {
  transportation: number;
  electricity: number;
  food: number;
  shopping: number;
  total: number;
}

export interface CarbonResult {
  totalAnnualKgCo2: number;
  breakdown: EmissionBreakdown;
  /** 0-100 percentile vs global average (lower = better) */
  percentileVsGlobal: number;
  /** Score from 0-100 (higher = better) */
  carbonScore: number;
}

export interface CarbonProfile {
  id: number;
  userId: number;
  inputs: CarbonInput;
  result: CarbonResult;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardMetrics {
  carbonScore: number;
  totalAnnualKgCo2: number;
  breakdown: EmissionBreakdown;
  /** Change vs last month in kg CO2 */
  monthlyDelta: number;
  goalsCount: number;
  goalsCompleted: number;
  totalCo2Saved: number;
}

export interface EmissionHistoryPoint {
  month: string;
  year: number;
  totalKgCo2: number;
  score: number;
}

export type GoalInputCategory = typeof GoalInputCategory[keyof typeof GoalInputCategory];


export const GoalInputCategory = {
  transportation: 'transportation',
  electricity: 'electricity',
  food: 'food',
  shopping: 'shopping',
  other: 'other',
} as const;

export interface GoalInput {
  title: string;
  /** @nullable */
  description?: string | null;
  targetKgCo2Reduction: number;
  category: GoalInputCategory;
  deadline: string;
}

export interface GoalUpdate {
  currentProgress?: number;
  completed?: boolean;
}

export interface Goal {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  targetKgCo2Reduction: number;
  currentProgress: number;
  category: string;
  deadline: string;
  completed: boolean;
  createdAt: string;
}

export type SimulationInputDietShift = typeof SimulationInputDietShift[keyof typeof SimulationInputDietShift];


export const SimulationInputDietShift = {
  none: 'none',
  vegetarian: 'vegetarian',
  vegan: 'vegan',
} as const;

export interface SimulationInput {
  /** 0-100 percent adoption of public transport */
  publicTransportAdoptionPct: number;
  /** 0-100 percent renewable energy */
  renewableEnergyPct: number;
  dietShift: SimulationInputDietShift;
  /** 0-100 percent reduction in flights */
  flightReductionPct: number;
}

export interface SimulationResult {
  currentAnnualKgCo2: number;
  projectedAnnualKgCo2: number;
  deltaKgCo2: number;
  percentageReduction: number;
  /** Number of trees needed to absorb this amount in a year */
  treesEquivalent: number;
  carFreeDays: number;
  explanation: string;
  breakdown: EmissionBreakdown;
  projectedBreakdown: EmissionBreakdown;
}

export type RecommendationCostDirection = typeof RecommendationCostDirection[keyof typeof RecommendationCostDirection];


export const RecommendationCostDirection = {
  save: 'save',
  cost: 'cost',
} as const;

export type RecommendationDifficulty = typeof RecommendationDifficulty[keyof typeof RecommendationDifficulty];


export const RecommendationDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type RecommendationCategory = typeof RecommendationCategory[keyof typeof RecommendationCategory];


export const RecommendationCategory = {
  transportation: 'transportation',
  electricity: 'electricity',
  food: 'food',
  shopping: 'shopping',
  other: 'other',
} as const;

export interface Recommendation {
  id: string;
  action: string;
  description?: string;
  estimatedKgCo2Reduction: number;
  /** Annual cost impact in GBP (positive = savings) */
  costImpact: number;
  costDirection: RecommendationCostDirection;
  difficulty: RecommendationDifficulty;
  category: RecommendationCategory;
  priority: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  carbonScore: number;
  scoreImprovement: number;
  totalCo2Saved: number;
}

export interface Challenge {
  id: number;
  title: string;
  description: string;
  participantCount: number;
  targetKgCo2: number;
  category: string;
  endDate: string;
  active: boolean;
}

export interface CommunityStats {
  totalUsers: number;
  totalCo2Saved: number;
  activeChallenges: number;
  avgCarbonScore: number;
}