# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zerofy API — AI Climate Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Challenge {
  id: number;
  title: string;
  description: string;
  participantCount: number;
  targetKgCo2: number;
  category: string;
  endDate: string;
  active: boolean;
}