# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zerofy API — AI Climate Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { GoalInputCategory } from './goalInputCategory';

export interface GoalInput {
  title: string;
  /** @nullable */
  description?: string | null;
  targetKgCo2Reduction: number;
  category: GoalInputCategory;
  deadline: Date;
}