# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zerofy API — AI Climate Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationCategory } from './recommendationCategory';
import type { RecommendationCostDirection } from './recommendationCostDirection';
import type { RecommendationDifficulty } from './recommendationDifficulty';

export interface Recommendation {
  id: string;
  action: string;
  description?: string;
  estimatedKgCo2Reduction: number;
  /** Annual cost impact in GBP (positive = savings) */
  costImpact: number;
  costDirection: RecommendationCostDirection;
  difficulty: RecommendationDifficulty;
  category: RecommendationCategory;
  priority: number;
}