# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zerofy API — AI Climate Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationInputDietShift } from './simulationInputDietShift';

export interface SimulationInput {
  /** 0-100 percent adoption of public transport */
  publicTransportAdoptionPct: number;
  /** 0-100 percent renewable energy */
  renewableEnergyPct: number;
  dietShift: SimulationInputDietShift;
  /** 0-100 percent reduction in flights */
  flightReductionPct: number;
}